requirements.txt
tensorflow
keras
numpy
tqdm
PyYAML
mlflow
pillow
-e .